Move algod dockerfile to new directory and refactor it
Now it is more configurable to improve the dev workflow
